\docType{data}
\name{exampleDataset}
\alias{exampleDataset}
\title{Example Dataset}
\description{
Dataset containing the information of 30 ballet dancers
}
\details{
The variables included in the dataset are: \itemize{
\item\code{weight} A vector with the weight in kilos of
each dancer \item\code{turns} Number of turns performed by
each dancer with a single impulse \item\code{sex} Sex of
the dancer 1 female 0 male \item\code{height} A vector with
the height in centimeters of each dancer
\item\code{training} Years of ballet training per dancer }
}

