\docType{package}
\name{RPackageProblemSet}
\alias{RPackageProblemSet}
\alias{RPackageProblemSet-package}
\title{RPackageProblemSet}
\description{
The RPackageProblemSet package performs regressions.
}
\examples{
\dontrun{
demo(reg.vars)
demo(sig.store)
}
}
\author{
Michelle Torres \email{smtorres@wustl.edu} and
}

